fix: Corregir autenticación y migración completa a PostgreSQL
- Actualizar entidades Usuario y Rol para PostgreSQL
  - Mapear nombres de columnas correctamente (id_usuario, nombre, apellido, contrasena, etc.)
  - Cambiar tipos de datos: tinyint->boolean, datetime->timestamp, double->decimal

- Corregir servicios de autenticación
  - Actualizar AuthService para usar campo 'estado' en lugar de 'activo'
  - Corregir jwt.strategy.ts para verificar campo 'estado'
  - Reescribir findOneWithPassword usando QueryBuilder correcto

- Actualizar DTOs para mayor flexibilidad
  - Hacer campos opcionales en CreateUsuarioDto
  - Agregar soporte para campo 'estado'

- Corregir configuración TypeORM
  - Cambiar tipo de base de datos de 'mysql' a 'postgres' en app.module.ts

- Eliminar referencias a archivos CSS eliminados
  - Remover imports de Estilos-Home.css y Estilos-Inicio.css

- Login y registro funcionando correctamente con PostgreSQL

+++ src/modules/carrito-producto/entities/carrito-producto.entity.ts
@@ -14,7 +14,7 @@ export class CarritoProducto {
   cantidad: number;
 
   @ApiProperty({ description: 'Precio unitario', example: 50000 })
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'decimal', nullable: true })
   precio: number;
 
   @ApiProperty({ description: 'ID del carrito' })
