+++ src/modules/servicios/entities/servicio.entity.ts
@@ -21,7 +21,7 @@ export class Servicio {
   nombre: string;
 
   @ApiProperty({ description: 'Precio del servicio', example: 50000 })
-  @Column({ type: 'double', nullable: true })
+  @Column({ type: 'decimal', nullable: true })
   precio: number;
 
   @ApiProperty({ description: 'ID de la veterinaria' })
